Remove a few more `BoardRunner.load` call sites and simplify it too,
